+++ Reverser.h
@@ -7,7 +7,7 @@
 class Reverser {
  private:
   int numDigits(int value);
-  std::string reverseStringExtra(std::string characters, int index);
+  std::string reverseStringExtra(std::string characters, size_t index);
 
  public:
   int reverseDigit(int value);
